requirements.txt
# Core dependencies
python-dotenv>=1.0.1
colorama>=0.4.4
pydantic>=2.6.3
typing-extensions>=4.9.0

# AI/ML Libraries
groq>=0.4.0
langchain>=0.1.9
langchain-core>=0.1.27
langchain-community>=0.0.24
langchain-groq>=0.0.1
langchain-huggingface>=0.0.11
langchain-chroma>=0.0.10
transformers>=4.46.3
sentence-transformers>=2.5.1
redis-om>=0.2.1
tiktoken>=0.6.0

# Vector Store and Embeddings
chromadb>=0.4.24
hnswlib>=0.7.0
faiss-cpu>=1.7.4
pysqlite3-binary>=0.5.2

# Document Processing
unstructured>=0.12.4
pypdf>=4.1.0
python-docx>=1.1.0
markdown>=3.5.2
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system == 'Windows'
tabulate>=0.9.0
pdf2image>=1.16.3
pdfminer.six>=20221105
docx2txt>=0.8

# Networking
aiohttp>=3.9.3
beautifulsoup4>=4.9.3
python-multipart>=0.0.9
requests>=2.31.0
urllib3>=2.2.0

# Frontend
streamlit>=1.31.0
streamlit-extras>=0.4.0
streamlit-chat>=0.1.1
Pillow>=10.4.0
plotly>=5.18.0

# Machine Learning
torch>=2.0.0
tensorflow>=2.18.0
tf-keras>=2.18.0
numpy>=1.26.0
pandas>=1.3.0
scikit-learn>=1.5.2

# Data Management
redis>=5.0.2
sqlalchemy>=2.0.0
pymongo>=4.6.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2
fsspec>=2024.10.0
PyYAML>=6.0.1
ujson>=5.9.0
python-jose[cryptography]>=3.3.0
uuid>=1.30
watchdog>=3.0.0

# Development Tools
black>=24.2.0
isort>=5.13.0
mypy>=1.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0